MMT-3495: Addressing feedback
