+++ static/src/js/constants/typeParamToHumanizedNameMap.js
@@ -1,0 +1,8 @@
+const typeParamToHumanizedNameMap = {
+  collections: 'collection',
+  services: 'service',
+  tools: 'tool',
+  variables: 'variable'
+}
+
+export default typeParamToHumanizedNameMap
